Renamed NetworkDefines.h --> NetworkDefinitions.h
